pyproject.toml
[tool.poetry]
name = "buecherei_srh_django"
version = "0.1.0"
description = ""
authors = ["Krytos <kmeinon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
django-bootstrap-modal-forms = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
